+++ interactions/api/http/http.pyi
