+++ interactions/api/models/channel.pyi
@@ -6,7 +6,7 @@ from .message import Message, Embed, MessageInteraction
 from ...models.component import ActionRow, Button, SelectMenu
 from .misc import DictSerializerMixin, Overwrite, Snowflake, MISSING
 from .user import User
-from ..http.HTTPClient import HTTPClient
+from ..http.http import HTTPClient
 
 class ChannelType(IntEnum):
     GUILD_TEXT: int
