+++ interactions/api/models/member.pyi
@@ -5,7 +5,7 @@ from .misc import DictSerializerMixin, MISSING, Snowflake
 from .role import Role
 from .user import User
 from .flags import Permissions
-from ..http.HTTPClient import HTTPClient
+from ..http.http import HTTPClient
 from .message import Message, Embed, MessageInteraction
 from ...models.component import ActionRow, Button, SelectMenu
